event_manager/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
crispy-bootstrap5==2024.10
    # via -r requirements.in
django==5.1.1
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-crispy-forms
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-crispy-forms==2.3
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-environ==0.11.2
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.27.2
    # via -r requirements.in
drf-spectacular-sidecar==2024.7.1
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.1
    # via django
tzdata==2024.2
    # via django
uritemplate==4.1.1
    # via drf-spectacular
whitenoise==6.7.0
    # via -r requirements.in
